Reorg utils a bit
