+++ javascript/core/src/engine/util/expressions.ts
@@ -5,17 +5,19 @@
 import { escape, op } from 'arquero'
 import type { Op } from 'arquero/dist/types/op/op'
 
-import type { Criterion } from '../../types.js'
-import {
+import type {
 	BooleanComparisonOperator,
+	Criterion,
+	NumericComparisonOperator,
+	StringComparisonOperator,
+} from '../../types.js'
+import {
 	BooleanLogicalOperator,
 	FieldAggregateOperation,
 	FilterCompareType,
-	NumericComparisonOperator,
-	StringComparisonOperator,
 	WindowFunction,
 } from '../../types.js'
-import { bool } from './data-types.js'
+import { compareValues } from './compare.js'
 import type { CompareWrapper } from './types.js'
 
 export function compareAll(
@@ -32,7 +34,7 @@ export function compareAll(
 			const { value, operator, type } = filter
 			const right =
 				type === FilterCompareType.Column ? d[`${value.toString()}`]! : value
-			return compareFunction(left, right, operator)
+			return compareValues(left, right, operator)
 		})
 		return evaluateBoolean(comparisons, logical)
 	}) as CompareWrapper
@@ -124,125 +126,8 @@ export function compare(
 		const right =
 			type === FilterCompareType.Column ? d[`${value.toString()}`]! : value
 
-		return compareFunction(left, right, operator)
+		return compareValues(left, right, operator)
 	}) as CompareWrapper
-}
-
-function compareFunction(
-	left: string | number | boolean,
-	right: string | number | boolean,
-	operator:
-		| NumericComparisonOperator
-		| StringComparisonOperator
-		| BooleanComparisonOperator,
-): 0 | 1 {
-	// start with the empty operators, because typeof won't work...
-	if (
-		operator === NumericComparisonOperator.IsEmpty ||
-		operator === StringComparisonOperator.IsEmpty ||
-		operator === BooleanComparisonOperator.IsEmpty
-	) {
-		return isEmpty(left)
-	} else if (
-		operator === NumericComparisonOperator.IsNotEmpty ||
-		operator === StringComparisonOperator.IsNotEmpty ||
-		operator === BooleanComparisonOperator.IsNotEmpty
-	) {
-		const empty = isEmpty(left)
-		return empty === 1 ? 0 : 1
-	} else if (typeof left === 'number') {
-		const num = +right
-		return compareValues(left, num, operator as NumericComparisonOperator)
-	} else if (typeof left === 'string') {
-		return compareStrings(
-			left,
-			`${right}`,
-			operator as StringComparisonOperator,
-		)
-	} else if (typeof left === 'boolean') {
-		const r = bool(right)
-		return compareBooleans(left, r, operator as BooleanComparisonOperator)
-	}
-	return 0
-}
-
-function isEmpty(value: string | number | boolean) {
-	if (value === null || value === undefined) {
-		return 1
-	}
-	if (typeof value === 'number' && isNaN(value)) {
-		return 1
-	}
-	if (typeof value === 'string' && value.length === 0) {
-		return 1
-	}
-	return 0
-}
-
-function compareStrings(
-	left: string,
-	right: string,
-	operator: StringComparisonOperator,
-): 1 | 0 {
-	const ll = left.toLocaleLowerCase()
-	const rl = right.toLocaleLowerCase()
-	switch (operator) {
-		case StringComparisonOperator.Contains:
-		case StringComparisonOperator.RegularExpression:
-			return op.match(ll, new RegExp(rl, 'gi'), 0) ? 1 : 0
-		case StringComparisonOperator.EndsWith:
-			return op.endswith(ll, rl, ll.length) ? 1 : 0
-		case StringComparisonOperator.Equals:
-			return ll.localeCompare(rl) === 0 ? 1 : 0
-		case StringComparisonOperator.NotEqual:
-			return ll.localeCompare(rl) !== 0 ? 1 : 0
-		case StringComparisonOperator.StartsWith:
-			return op.startswith(ll, rl, 0) ? 1 : 0
-		default:
-			throw new Error(`Unsupported string comparison operator: [${operator}]`)
-	}
-}
-
-function compareValues(
-	left: number,
-	right: number,
-	operator: NumericComparisonOperator,
-): 1 | 0 {
-	switch (operator) {
-		case NumericComparisonOperator.Equals:
-			return left === right ? 1 : 0
-		case NumericComparisonOperator.NotEqual:
-			return left !== right ? 1 : 0
-		case NumericComparisonOperator.GreaterThanOrEqual:
-			return left >= right ? 1 : 0
-		case NumericComparisonOperator.LessThanOrEqual:
-			return left <= right ? 1 : 0
-		case NumericComparisonOperator.GreaterThan:
-			return left > right ? 1 : 0
-		case NumericComparisonOperator.LessThan:
-			return left < right ? 1 : 0
-		default:
-			throw new Error(`Unsupported numeric comparison operator: [${operator}]`)
-	}
-}
-
-function compareBooleans(
-	left: boolean,
-	right: boolean,
-	operator: BooleanComparisonOperator,
-): 1 | 0 {
-	switch (operator) {
-		case BooleanComparisonOperator.Equals:
-			return left === right ? 1 : 0
-		case BooleanComparisonOperator.NotEqual:
-			return left !== right ? 1 : 0
-		case BooleanComparisonOperator.IsTrue:
-			return left === true ? 1 : 0
-		case BooleanComparisonOperator.IsFalse:
-			return left === false ? 1 : 0
-		default:
-			throw new Error(`Unsupported boolean comparison operator: [${operator}]`)
-	}
 }
 
 const fieldOps = new Set([
@@ -262,49 +147,4 @@ export function singleExpression(
 		)
 	}
 	return op[operation](column)
-}
-
-export function fixedBinCount(
-	column: string,
-	min: number,
-	max: number,
-	count: number,
-	clamped?: boolean,
-	distinct?: number,
-): string | object {
-	const step = (max - min) / count
-	return fixedBinStep(column, min, max, step, clamped, distinct)
-}
-
-export function fixedBinStep(
-	column: string,
-	min: number,
-	max: number,
-	step: number,
-	clamped?: boolean,
-	distinct?: number,
-): string | object {
-	const count = Math.ceil((max - min) / step)
-	const ultimate = min + step * count
-	const penultimate = min + step * (count - 1)
-	// if the ultimate bin is >= the max, put those values in the prior bin
-	// this is due to arquero's exclusive max bound, which will just bin those exact
-	// matches into the final bin, disrupting the expected bin count by adding one
-	const rebinmax = ultimate >= max
-	const nobin = distinct !== undefined && distinct <= count
-	return escape((d: any) => {
-		const value = d[column]
-		if (nobin) {
-			return value
-		}
-		const candidate = op.bin(value, min, max, step)
-		if (clamped) {
-			if (candidate === -Infinity) {
-				return min
-			} else if (candidate === Infinity) {
-				return rebinmax ? penultimate : ultimate
-			}
-		}
-		return candidate === max && rebinmax ? penultimate : candidate
-	})
 }
