+++ javascript/core/src/engine/util/index.ts
@@ -2,5 +2,6 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+export * from './bins.js'
 export * from './expressions.js'
 export * from './sets.js'
